requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.3
django-allauth==0.33.0
django-appconf==1.0.2
django-braces==1.11.0
django-ckeditor==5.3.0
django-imagekit==4.0.1
django-js-asset==0.1.1
django-model-utils==3.0.0
django-mptt==0.8.7
django-notifications-hq==1.3
idna==2.6
jsonfield==2.0.2
Markdown==2.6.8
oauthlib==2.0.2
olefile==0.44
pilkit==2.0
Pillow==4.2.1
Pygments==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
timeago==1.0.7
urllib3==1.22
